+++ EVTCAnalytics/Processing/LogProcessor.cs
@@ -1890,6 +1890,39 @@ namespace GW2Scratch.EVTCAnalytics.Processing
 
 						return new AgentGadgetNameEvent(item.Time, GetAgentByAddress(item.SrcAgent), item.DstAgent);
 					}
+					case StateChange.MissileEffect:
+					{
+						// dst_agent: owner of missile
+						// skillid: effect id
+						// value: *(uint32_t*)&value duration
+						// pad61: (uint32_t*)&pad61 is uint32[1], trackable id
+
+						return new MissileEffectEvent(item.Time, GetAgentByAddress(item.DstAgent), item.SkillId, item.Value, item.Padding);
+					}
+					case StateChange.GadgetCaptureOutlineShow:
+					{
+						// src_agent: relates to agent
+						// buff: wrbg colour
+
+						return new GadgetCaptureOutlineShowEvent(item.Time, GetAgentByAddress(item.SrcAgent), item.Buff);
+					}
+					case StateChange.GadgetCaptureSplitPercent:
+					{
+						// src_agent: relates to agent
+						// value: *(float*)&ev->value percent
+						// buff: wrbg capping from
+						// result: wrbg capping by
+
+						// Note by deltaconnected on Discord: value is the float without any conversion or lower-precision-inting
+
+						return new GadgetCaptureSplitPercentEvent(item.Time, GetAgentByAddress(item.SrcAgent), item.Value, item.Buff, (byte)item.Result);
+					}
+					case StateChange.GadgetCaptureOutlineHide:
+					{
+						// src_agent: relates to agent
+
+						return new GadgetCaptureOutlineHideEvent(item.Time, GetAgentByAddress(item.SrcAgent));
+					}
 					default:
 						return new UnknownEvent(item.Time, item);
 				}